Pass RegionalPromptingExtension down to the CustomDoubleStreamBlockProcessor in FLUX.
